Support multiple URLs in a single analysis
This allows multiple URLs to be passed and the results combined into a
single table.

This did run into an issue to oclif whereby passing an arbitrary number
of arguments doesn't seem to work correctly. It does support multiple
arguments for a flag though so this will suffice for the moment. Beware
that this behaviour may change as there is a Github issue open as it has
caused some other problems.

`https://github.com/oclif/oclif/issues/190`